made separate components for the dashboard
